async project exec

+++ src/packages/backend/execute-code.ts
@@ -6,30 +6,52 @@
 // Execute code in a subprocess.
 
 import { callback } from "awaiting";
-import { spawn } from "node:child_process";
+import LRU from "lru-cache";
+import {
+  ChildProcessWithoutNullStreams,
+  spawn,
+  SpawnOptionsWithoutStdio,
+} from "node:child_process";
 import { chmod, mkdtemp, rm, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import shellEscape from "shell-escape";
 
+import { envToInt } from "@cocalc/backend/misc/env-to-number";
 import getLogger from "@cocalc/backend/logger";
 import { aggregate } from "@cocalc/util/aggregate";
 import { callback_opts } from "@cocalc/util/async-utils";
-import { to_json, trunc, walltime } from "@cocalc/util/misc";
+import { to_json, trunc, uuid, walltime } from "@cocalc/util/misc";
 import { envForSpawn } from "./misc";
 
-import type {
-  ExecuteCodeFunctionWithCallback,
-  ExecuteCodeOptions,
-  ExecuteCodeOptionsWithCallback,
-  ExecuteCodeOutput,
+import {
+  ExecuteCodeOutputAsync,
+  ExecuteCodeOutputBlocking,
+  isExecuteCodeOptionsAsyncGet,
+  type ExecuteCodeFunctionWithCallback,
+  type ExecuteCodeOptions,
+  type ExecuteCodeOptionsAsyncGet,
+  type ExecuteCodeOptionsWithCallback,
+  type ExecuteCodeOutput,
 } from "@cocalc/util/types/execute-code";
 
 const log = getLogger("execute-code");
 
+const ASYNC_CACHE_MAX = envToInt("COCALC_PROJECT_ASYNC_EXEC_CACHE_MAX", 100);
+const ASYNC_CACHE_TTL_S = envToInt("COCALC_PROJECT_ASYNC_EXEC_TTL_S", 60 * 60);
+
+const asyncCache = new LRU<string, ExecuteCodeOutputAsync>({
+  max: ASYNC_CACHE_MAX,
+  ttl: 1000 * ASYNC_CACHE_TTL_S,
+  ttlAutopurge: true,
+  allowStale: true,
+  updateAgeOnGet: true,
+  updateAgeOnHas: true,
+});
+
 // Async/await interface to executing code.
 export async function executeCode(
-  opts: ExecuteCodeOptions,
+  opts: ExecuteCodeOptions | ExecuteCodeOptionsAsyncGet,
 ): Promise<ExecuteCodeOutput> {
   return await callback_opts(execute_code)(opts);
 }
@@ -48,20 +70,34 @@ export const execute_code: ExecuteCodeFunctionWithCallback = aggregate(
   },
 );
 
+async function clean_up_tmp(tempDir: string | undefined) {
+  if (tempDir) {
+    await rm(tempDir, { force: true, recursive: true });
+  }
+}
+
 // actual implementation, without the aggregate wrapper
 async function executeCodeNoAggregate(
-  opts: ExecuteCodeOptions,
+  opts: ExecuteCodeOptions | ExecuteCodeOptionsAsyncGet,
 ): Promise<ExecuteCodeOutput> {
-  if (opts.args == null) opts.args = [];
-  if (opts.timeout == null) opts.timeout = 10;
-  if (opts.ulimit_timeout == null) opts.ulimit_timeout = true;
-  if (opts.err_on_exit == null) opts.err_on_exit = true;
-  if (opts.verbose == null) opts.verbose = true;
+  if (isExecuteCodeOptionsAsyncGet(opts)) {
+    const cached = asyncCache.get(opts.async_get);
+    if (cached != null) {
+      return cached;
+    } else {
+      throw new Error(`Async operation '${opts.async_get}' does not exist.`);
+    }
+  }
+
+  opts.args ??= [];
+  opts.timeout ??= 10;
+  opts.ulimit_timeout ??= true;
+  opts.err_on_exit ??= true;
+  opts.verbose ??= true;
 
   if (opts.verbose) {
     log.debug(`input: ${opts.command} ${opts.args?.join(" ")}`);
   }
-
   const s = opts.command.split(/\s+/g); // split on whitespace
   if (opts.args?.length === 0 && s.length > 1) {
     opts.bash = true;
@@ -115,16 +151,90 @@ async function executeCodeNoAggregate(
       await writeFile(tempPath, cmd);
       await chmod(tempPath, 0o700);
     }
-    return await callback(doSpawn, { ...opts, origCommand });
-  } finally {
-    // clean up
-    if (tempDir) {
-      await rm(tempDir, { force: true, recursive: true });
+
+    if (opts.async_call) {
+      // we return an ID, the caller can then use it to query the status
+      opts.max_output ??= 1024 * 1024; // we limit how much we keep in memory, to avoid problems;
+      opts.timeout ??= 10 * 60;
+      const job_id = uuid();
+      const start = new Date();
+      const started: ExecuteCodeOutputAsync = {
+        type: "async",
+        stdout: `Process started running at ${start.toISOString()}`,
+        stderr: "",
+        exit_code: 0,
+        start: start.getTime(),
+        job_id,
+        status: "running",
+      };
+      asyncCache.set(job_id, started);
+
+      doSpawn({ ...opts, origCommand, job_id }, async (err, result) => {
+        try {
+          const started = asyncCache.get(job_id)?.start ?? 0;
+          const info: Omit<
+            ExecuteCodeOutputAsync,
+            "stdout" | "stderr" | "exit_code"
+          > = {
+            job_id,
+            type: "async",
+            elapsed_s: (Date.now() - started) / 1000,
+            start: start.getTime(),
+            status: "error",
+          };
+          if (err) {
+            asyncCache.set(job_id, {
+              stdout: "",
+              stderr: `${err}`,
+              exit_code: 1,
+              ...info,
+            });
+          } else if (result != null) {
+            asyncCache.set(job_id, {
+              ...result,
+              ...info,
+              ...{ status: "completed" },
+            });
+          } else {
+            asyncCache.set(job_id, {
+              stdout: "",
+              stderr: `No result`,
+              exit_code: 1,
+              ...info,
+            });
+          }
+        } finally {
+          await clean_up_tmp(tempDir);
+        }
+      });
+
+      return started;
+    } else {
+      // This is the blocking variant
+      return await callback(doSpawn, { ...opts, origCommand });
     }
+  } finally {
+    // do not delete the tempDir in async mode!
+    if (!opts.async_call) await clean_up_tmp(tempDir);
   }
 }
 
-function doSpawn(opts, cb) {
+function update_async(
+  job_id: string | undefined,
+  stream: "stdout" | "stderr",
+  data: string,
+) {
+  if (!job_id) return;
+  const obj = asyncCache.get(job_id);
+  if (obj != null) {
+    obj[stream] = data;
+  }
+}
+
+function doSpawn(
+  opts,
+  cb: (err: string | undefined, result?: ExecuteCodeOutputBlocking) => void,
+) {
   const start_time = walltime();
 
   if (opts.verbose) {
@@ -138,7 +248,7 @@ function doSpawn(opts, cb) {
       "seconds",
     );
   }
-  const spawnOptions = {
+  const spawnOptions: SpawnOptionsWithoutStdio = {
     detached: true, // so we can kill the entire process group if it times out
     cwd: opts.path,
     ...(opts.uid ? { uid: opts.uid } : undefined),
@@ -150,8 +260,8 @@ function doSpawn(opts, cb) {
     },
   };
 
-  let r,
-    ran_code = false;
+  let r: ChildProcessWithoutNullStreams;
+  let ran_code = false;
   try {
     r = spawn(opts.command, opts.args, spawnOptions);
     if (r.stdout == null || r.stderr == null) {
@@ -187,6 +297,7 @@ function doSpawn(opts, cb) {
     } else {
       stdout += data;
     }
+    update_async(opts.job_id, "stdout", stdout);
   });
 
   r.stderr.on("data", (data) => {
@@ -198,6 +309,7 @@ function doSpawn(opts, cb) {
     } else {
       stderr += data;
     }
+    update_async(opts.job_id, "stderr", stderr);
   });
 
   let stderr_is_done = false;
@@ -215,7 +327,7 @@ function doSpawn(opts, cb) {
   });
 
   r.on("exit", (code) => {
-    exit_code = code;
+    exit_code = code != null ? code : undefined;
     finish();
   });
 
@@ -270,12 +382,17 @@ function doSpawn(opts, cb) {
       const x = opts.origCommand
         ? opts.origCommand
         : `'${opts.command}' (args=${opts.args?.join(" ")})`;
-      cb(
-        `command '${x}' exited with nonzero code ${exit_code} -- stderr='${trunc(
-          stderr,
-          1024,
-        )}'`,
-      );
+      if (opts.job_id) {
+        cb(stderr);
+      } else {
+        // sync behavor, like it was before
+        cb(
+          `command '${x}' exited with nonzero code ${exit_code} -- stderr='${trunc(
+            stderr,
+            1024,
+          )}'`,
+        );
+      }
     } else if (!ran_code) {
       // regardless of opts.err_on_exit !
       const x = opts.origCommand
@@ -297,7 +414,7 @@ function doSpawn(opts, cb) {
         // if exit-code not set, may have been SIGKILL so we set it to 1
         exit_code = 1;
       }
-      cb(undefined, { stdout, stderr, exit_code });
+      cb(undefined, { type: "blocking", stdout, stderr, exit_code });
     }
   };
 
@@ -317,8 +434,10 @@ function doSpawn(opts, cb) {
         );
       }
       try {
-        killed = true;
-        process.kill(-r.pid, "SIGKILL"); // this should kill process group
+        killed = true; // we set the kill flag in any case – i.e. process will no longer exist
+        if (r.pid != null) {
+          process.kill(-r.pid, "SIGKILL"); // this should kill process group
+        }
       } catch (err) {
         // Exceptions can happen, which left uncaught messes up calling code big time.
         if (opts.verbose) {
